# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Symphony Orchestrator API
  version: 1.0.0
  description: Symphony Orchestrator API for Vertex Workbench provisioning and quota management
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects/{projectId}/workbench/instances:
    get:
      tags: [Workbench]
      summary: List Workbench instances visible to the caller
      operationId: listWorkbenchInstances
      parameters:
        - $ref: "#/components/parameters/projectId"
      responses:
        "200":
          description: Instances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkbenchInstance" }
        "401":
          $ref: "#/components/responses/Unauthorized"
      security:
        - bearerAuth: [workbench.user]

    post:
      tags: [Workbench]
      summary: Create a Workbench instance (hard quota enforced; system cap max 3 instances)
      operationId: createWorkbenchInstance
      parameters:
        - $ref: "#/components/parameters/projectId"
        - $ref: "#/components/parameters/groupId"
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInstanceRequest" }
      responses:
        "202":
          description: Provisioning started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AsyncOperation" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Name conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Rejected due to quota/policy violation (hard limit)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuotaViolationError" }
      security:
        - bearerAuth: [workbench.user]
  /v1/projects/{projectId}/workbench/instances/{instanceId}:
    get:
      tags: [Workbench]
      summary: Get a Workbench instance
      operationId: getWorkbenchInstance
      parameters:
        - $ref: "#/components/parameters/projectId"
        - name: instanceId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Instance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkbenchInstance" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - bearerAuth: [workbench.user]

    delete:
      tags: [Workbench]
      summary: Delete a Workbench instance
      operationId: deleteWorkbenchInstance
      parameters:
        - $ref: "#/components/parameters/projectId"
        - name: instanceId
          in: path
          required: true
          schema: { type: string }
      responses:
        "202":
          description: Deletion started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AsyncOperation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - bearerAuth: [workbench.user]
  /v1/projects/{projectId}/workbench/instances/{instanceId}:start:
    post:
      tags: [Workbench]
      summary: Start a stopped Workbench instance
      operationId: startWorkbenchInstance
      parameters:
        - $ref: "#/components/parameters/projectId"
        - name: instanceId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "202":
          description: Start initiated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AsyncOperation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Instance is not in a startable state (e.g. already RUNNING)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.user]
  /v1/projects/{projectId}/workbench/instances/{instanceId}:stop:
    post:
      tags: [Workbench]
      summary: Stop a running Workbench instance
      operationId: stopWorkbenchInstance
      parameters:
        - $ref: "#/components/parameters/projectId"
        - name: instanceId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/idempotencyKey"
      responses:
        "202":
          description: Stop initiated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AsyncOperation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Instance is not in a stoppable state (e.g. already STOPPED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
      security:
        - bearerAuth: [workbench.user]

  # -------------------------
  # Usage (User)
  # -------------------------
...
  /v1/operations/{operationId}:
    get:
      tags: [Operations]
      summary: Get operation status
      operationId: getOperation
      description: |
        Accessible to any authenticated user or manager. Server-side access control ensures
        callers can only poll operations they initiated.
      parameters:
        - name: operationId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Operation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AsyncOperation" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller did not initiate this operation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - bearerAuth: [workbench.user]
components:
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
    groupId:
      name: groupId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/GroupId" }
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      # NOTE: OpenAPI scopes (e.g. bearerAuth: [workbench.manager]) are only enforced
      # for oauth2/openIdConnect scheme types. For http bearer, the scope array is
      # ignored by validators and tooling — it is used here as intent documentation only.
      # Actual scope enforcement is performed server-side via Spring Security @PreAuthorize.
  schemas:
...
    BankUserId:
      type: string
      description: 1bankid used in Workbench labels.user
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._-]+$'
    GroupId:
      type: string
      description: Unique directory group identifier (GUID/DN)
      minLength: 1
      maxLength: 128
...
    SystemCaps:
      type: object
      required: [maxInstancesPerUser]
      properties:
        maxInstancesPerUser:
          type: integer
          const: 3
          description: Hard platform cap enforced regardless of manager quota.
    EffectiveLimits:
      type: object
      required: [maxInstances, maxTotalVcpu, allowedMachineTypes]
      properties:
        maxInstances:
          type: integer
          minimum: 0
          description: min(systemCaps.maxInstancesPerUser=3, manager/default quota maxInstances)
        maxTotalVcpu:
          type: integer
          minimum: 0
        allowedMachineTypes:
          type: array
          items: { type: string }
        maxBootDiskGb:
          type: integer
          minimum: 50
        allowedZones:
          type: array
          items: { type: string }
    UsageTotals:
      type: object
      required: [instances, totalVcpu]
      properties:
        instances: { type: integer, minimum: 0 }
        totalVcpu: { type: integer, minimum: 0 }
    UsageComputation:
      type: object
      required: [computedFrom, assetType, query]
      properties:
        computedFrom:
          type: string
          enum: [CLOUD_ASSET_INVENTORY]
        assetType:
          type: string
          const: compute.googleapis.com/Instance
        query:
          type: string
          description: CAI query used.
          example: "labels.notebooks-product:* AND labels.user:u12345"
    UserQuotaUsage:
      type: object
      required: [projectId, userId, computedAt, computation, systemCaps, effectiveLimits, usage, remaining]
      properties:
        projectId: { type: string }
        userId: { $ref: "#/components/schemas/BankUserId" }
        computedAt: { type: string, format: date-time }
        computation: { $ref: "#/components/schemas/UsageComputation" }
        systemCaps: { $ref: "#/components/schemas/SystemCaps" }
        effectiveLimits: { $ref: "#/components/schemas/EffectiveLimits" }
        usage: { $ref: "#/components/schemas/UsageTotals" }
        remaining: { $ref: "#/components/schemas/UsageTotals" }
...
    CreateInstanceRequest:
      type: object
      required: [displayName, zone, machineType]
      additionalProperties: false
      properties:
        displayName: { type: string }
        zone: { type: string }
        machineType: { type: string }
        bootDiskGb: { type: integer, minimum: 50 }
        labels:
          type: object
          additionalProperties: { type: string }
          description: Optional extra labels; server will set/overwrite labels.user=<caller 1bankid>.
    WorkbenchInstance:
      type: object
      required: [id, displayName, state, machineType, zone, ownerUserId, createdAt]
      properties:
        id: { type: string }
        displayName: { type: string }
        state: { type: string, enum: [PROVISIONING, RUNNING, STOPPED, DELETING, ERROR] }
        machineType: { type: string }
        zone: { type: string }
        ownerUserId: { $ref: "#/components/schemas/BankUserId" }
        createdAt: { type: string, format: date-time }

    # -------------------------
    # Operations & Errors
    # -------------------------
    AsyncOperation:
      type: object
      required: [id, status, createdAt]
      properties:
        id: { type: string }
        status: { type: string, enum: [PENDING, RUNNING, DONE, FAILED] }
        createdAt: { type: string, format: date-time }
        doneAt: { type: string, format: date-time }
        resourceId: { type: string }
        error:
          oneOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
    Violation:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - SYSTEM_MAX_INSTANCES_EXCEEDED
            - MAX_INSTANCES_EXCEEDED
            - MAX_VCPU_EXCEEDED
            - MACHINE_TYPE_NOT_ALLOWED
            - ZONE_NOT_ALLOWED
            - DISK_TOO_LARGE
        message: { type: string }
        details:
          type: object
          additionalProperties: true
          description: Optional structured context for this violation.
          properties:
            limit: { type: integer, description: The applicable limit that was exceeded }
            current: { type: integer, description: The caller's current value at the time of the check }
            requested: { type: integer, description: The value that was requested }
            value: { type: string, description: The specific disallowed value (e.g. machine type or zone name) }
    QuotaViolationError:
      type: object
      required: [error, violations, snapshot]
      properties:
        error:
          type: string
          enum: [QUOTA_VIOLATION]
        violations:
          type: array
          items: { $ref: "#/components/schemas/Violation" }
        snapshot:
          $ref: "#/components/schemas/UserQuotaUsage"
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code: { type: string }
        requestId: { type: string, description: Correlation ID for tracing this error in server logs }